Allow optionally ignoring sphinx directives/roles
Sphinx roles can create false alarms and false errors
if we do not ignore them, so this adds the ability to ignore
them (and turns it on by default).

Do note that when enabled this module will patch ignore
directives and roles into the sphinx code-base and then make
an attempt to remove them when it has finished (this is inherently
not thread-safe).